move local_storage_path to core and update imports
