+++ interpreter/core/core.py
@@ -12,13 +12,13 @@ from datetime import datetime
 
 from ..terminal_interface.terminal_interface import terminal_interface
 from ..terminal_interface.utils.display_markdown_message import display_markdown_message
-from ..terminal_interface.utils.local_storage_path import get_storage_path
 from ..terminal_interface.utils.oi_dir import oi_dir
 from .computer.computer import Computer
 from .default_system_message import default_system_message
 from .llm.llm import Llm
 from .respond import respond
 from .server import server
+from .utils.local_storage_path import get_storage_path
 from .utils.telemetry import send_telemetry
 from .utils.truncate_output import truncate_output
