+++ interpreter/terminal_interface/utils/ARCHIVE_get_config.py
@@ -8,7 +8,7 @@ from importlib import resources
 
 import yaml
 
-from .local_storage_path import get_storage_path
+from ...core.utils.local_storage_path import get_storage_path
 
 config_filename = "config.yaml"
